refactor: SM-205 Jango Context 삭제 및 재료 선택 모달 전달받는 prop 추가
