+++ src/components/domain/IngredientSelectModal/types.ts
@@ -1,5 +1,8 @@
+import type { IIngredient } from '@models';
+
 interface IngredientSelectModalProps {
   visible: boolean;
+  totalIngredientsList: { [key: string]: IIngredient };
   initialMainIngredient: number[];
   initialSubIngredient: number[];
   onClose(): void;
