--- src/contexts/Jango/types.ts
@@ -1,9 +1,0 @@
-import type { IIngredient } from '@models/types';
-
-interface IJangoContext {
-  totalIngredientsList: {
-    [key: string]: IIngredient;
-  };
-}
-
-export type { IJangoContext };
