+++ src/routes/JangoRoutes/index.tsx
@@ -2,19 +2,11 @@ import type { ReactElement } from 'react';
 import { Route } from 'react-router-dom';
 import PrivateRoutes from '../PrivateRoutes';
 import { JangoPage } from '@pages';
-import JangoProvider from '@contexts/Jango';
 
 const JangoRoutes = (): ReactElement => {
   return (
     <PrivateRoutes>
-      <Route
-        element={
-          <JangoProvider>
-            <JangoPage />
-          </JangoProvider>
-        }
-        path='/'
-      />
+      <Route element={<JangoPage />} path='/' />
     </PrivateRoutes>
   );
 };
